added config option for critical hit chances updated hit code updated version info updated orbit code to latest dev version

+++ src/eve-server/npc/NPCAI.h
@@ -99,7 +99,7 @@ protected:
     void AttackTarget(SystemEntity* pSE);
     void CheckDistance(SystemEntity* pSE);
 
-    double GetTargetTime();
+    float GetTargetTime();
 
     int8 m_state;
